api/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.0
pydantic==2.6.0
python-multipart==0.0.9
loguru==0.7.2
numpy==1.26.3
scipy==1.12.0
requests==2.31.0
python-dotenv==1.0.0
httpx==0.27.0
